Add SpeakerInitials component
